Bclerc

+++ back/src/wschat/wschat.service.ts
@@ -1,45 +1,74 @@
 import { Injectable } from '@nestjs/common';
 import { WebSocketServer } from '@nestjs/websockets';
-import { ChatRoom, Message, User } from '@prisma/client';
+import { ChatRoom, Message, PenaltyType, User } from '@prisma/client';
 import { ChatService } from 'src/chat/chat.service';
 import { SubscribeRoomDto } from 'src/chat/dto/subscribe-room.dto';
 import { ChatRoomI, MessageI, newChatRoomI } from 'src/chat/interfaces/chatRoom.interface';
 import { EjectRoomI } from 'src/chat/interfaces/eject-room-i.interface';
 import { DemoteUserI, PromoteUserI } from 'src/chat/interfaces/promote-user-i.interface';
+import { PasswordUtils } from 'src/chat/utils/chat-utils';
 import { OnlineUserService } from 'src/onlineusers/onlineuser.service';
 import { BasicUserI } from 'src/user/interface/basicUser.interface';
 import { UserService } from 'src/user/user.service';
+import { EventEmitter2 } from '@nestjs/event-emitter';
+import { PenaltiesService } from 'src/chat/services/penalties/penalties.service';
 
 @Injectable()
 export class WschatService {
+
+  passUtils: PasswordUtils = new PasswordUtils();
 
   @WebSocketServer() server;
 
   constructor(
     private chatService: ChatService,
     private userService: UserService,
-    private onlineUserService: OnlineUserService
+    private onlineUserService: OnlineUserService,
+    private penaltiesService: PenaltiesService,
+    private eventEmitter: EventEmitter2
+
   ) { }
 
   async initUser(user: BasicUserI) {
-
     let rooms: ChatRoomI[];
+  
     rooms = await this.chatService.getRoomsFromUser(user.id);
     this.sendToUser(user, 'rooms', rooms);
   }
 
   async subscribeToRoom(socketId: string, subRoom: SubscribeRoomDto) {
-
     const user = this.onlineUserService.getUser(socketId);
-    let room = await this.chatService.getRoomById(subRoom.roomId);
+    const room = await this.chatService.getRoomById(subRoom.roomId);
+    const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);
+    
+    
+    if (penalty && penalty.type === PenaltyType.BAN) {
+      this.eventEmitter.emit('room.user.join', {
+        room: room,
+        user: user,
+        success: false,
+        message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime)
+      });
+      return ;
+    }
 
     if (user) {
-
       if (!await this.chatService.canJoin(room.id, subRoom.password)) {
-        return this.sendToUser(user, 'notification', "Le mot de passe est incorrect");
+        this.eventEmitter.emit('room.user.join', {
+          room: room,
+          user: user,
+          success: false,
+          message: 'Mot de passe incorrect',
+         });
+         return ;
       }
+
       await this.chatService.addUsersToRoom(room.id, user.id);
-      this.updateRoomForUsersInRoom(room.id);
+      this.eventEmitter.emit('room.user.join', {
+        room: room,
+        user: user,
+        success: true,
+      });
     }
   }
 
@@ -47,13 +76,11 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const target = await this.userService.findOne(room.targetId);
     const roomToEject = await this.chatService.getRoomById(room.roomId);
-    if (user) {
+
+    if (user && target.id != roomToEject.ownerId) {
       if (roomToEject.admins.find(admin => admin.id == user.id)) {
         await this.chatService.removeUsersFromRoom(roomToEject.id, target.id);
-        await this.updateRoomForUsersInRoom(roomToEject.id);
-        await this.updateUserRooms(target);
-        this.sendToUsersInRoom(roomToEject.id, 'notification', target.intra_name + " a été ejecté du salon " + roomToEject.name);
-        this.sendToUser(target, 'notification', "Vous avez été ejecté du salon " + roomToEject.name);
+        this.eventEmitter.emit('room.user.kicked', { room: roomToEject, user: target, kicker: user });
       }
     }
   }
@@ -64,11 +91,15 @@ export class WschatService {
     const roomToPromote = await this.chatService.getRoomById(event.roomId);
 
     if (user) {
-      if (roomToPromote.admins.find(admin => admin.id == user.id)) {
+      if (user.id == roomToPromote.ownerId) {
         this.chatService.addAdminsToRoom(roomToPromote.id, target.id);
         this.updateRoomForUsersInRoom(roomToPromote.id);
-        this.sendToUsersInRoom(roomToPromote.id, 'notification', target.intra_name + " a été promu admin du salon " + roomToPromote.name);
-        this.sendToUser(target, 'notification', "Vous avez été promu admin du salon " + roomToPromote.name);
+        this.eventEmitter.emit('room.admin.update', {
+          isPromote: true,
+          user: target,
+          promoter: user,
+          room: roomToPromote
+        })
       }
     }
   }
@@ -79,16 +110,19 @@ export class WschatService {
     const roomToDemote = await this.chatService.getRoomById(event.roomId);
 
     if (user) {
-      if (roomToDemote.admins.find(admin => admin.id == user.id)) {
+      if (roomToDemote.ownerId == user.id && roomToDemote.ownerId != target.id) {
         this.chatService.removeAdminsFromRoom(roomToDemote.id, target.id);
         this.updateRoomForUsersInRoom(roomToDemote.id);
-        this.sendToUsersInRoom(roomToDemote.id, 'notification', target.intra_name + " a été dégradé du salon " + roomToDemote.name);
-        this.sendToUser(target, 'notification', "Vous avez été dégradé du salon " + roomToDemote.name);
+        this.eventEmitter.emit('room.admin.update', {
+          isPromote: false,
+          user: target,
+          promoter: user,
+          room: roomToDemote
+        })
       }
     }
   }
 
-  // A faire: Verifier sur le mec est mute dans le chat
   async newMessage(socketId: string, message: MessageI) {
     let messages: Message[];
     const user = this.onlineUserService.getUser(socketId);
@@ -96,7 +130,7 @@ export class WschatService {
 
     if (user) {
       await this.chatService.newMessage(message);
-      this.updateUsersMessagesInRoom(room);
+      this.eventEmitter.emit('room.message.new', { room: room });
     }
   }
 
@@ -106,8 +140,7 @@ export class WschatService {
     
     if (user) {
       newRoom = await this.chatService.createRoom(user, room);
-      this.updateRoomForUsersInRoom(newRoom.id);
-      this.sendToUsersInRoom(newRoom.id, 'notification', "Vous avez rejoint le salon " + newRoom.name);
+      this.eventEmitter.emit('room.new', { room: newRoom });
     }
   }
   
@@ -116,37 +149,29 @@ export class WschatService {
     const room = await this.chatService.getRoomById(roomId);
     
     if (user) {
-      await this.chatService.removeUsersFromRoom(room.id, user.id);
-      this.updateRoomForUsersInRoom(room.id);
-      this.updateUserRooms(user);
-      this.sendToUser(user, 'notification', "Vous avez quitté le salon " + room.name);
-    } else {
-      this.sendToUser(user, 'notification', "Une erreur s'est produite");
+      const newRoom = await this.chatService.removeUsersFromRoom(room.id, user.id);
+      if (room.ownerId == user.id) {
+        if (newRoom.users.length > 0) {
+          this.chatService.updateRoomOwner(room.id, newRoom.users[0].id);
+          this.updateRoomForUsersInRoom(room.id);
+        } else {
+          // delete
+          this.eventEmitter.emit('room.delete', { room: room });
+        }
+      }
+      this.eventEmitter.emit('room.user.leave', { room: room, user: user });
     }
   }
+
   async editRoom(socketId: string, newRoom: newChatRoomI){
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(newRoom.id);
     if (user) {
-      if (room.admins.find(admin => admin.id == user.id)) {
+      if (room.ownerId == user.id) {
         await this.chatService.editRoom(newRoom);
-        this.updateRoomForUsersInRoom(room.id);
-        this.sendToUsersInRoom(room.id, 'notification', "Le salon " + room.name + " a été modifié");
+        this.eventEmitter.emit('room.update', { room: newRoom });
       }
     }
-  }
-
-  async updateUserRooms(user: BasicUserI) {
-
-    const rooms = await this.chatService.getRoomsFromUser(user.id);
-    this.sendToUser(user, 'rooms', rooms);
-  }
-
-  async updateUsersMessagesInRoom (room: ChatRoomI) {
-      for (let user of room.users) {
-        const messages = await this.chatService.getMessagesFromRoomId(user.id, room.id);
-        this.sendToUser(user, 'messages', messages);
-      }
   }
 
   async updateRoomForUsersInRoom(roomId: number) {
@@ -160,14 +185,6 @@ export class WschatService {
 
   sendToUser(user: BasicUserI, prefix: string, data: any) {
     this.onlineUserService.sendToUser(user, prefix, data);
-  }
-
-  async sendToUsersInRoom(roomId: number, prefix: string, data: any) {
-    let room = await this.chatService.getRoomById(roomId);
-
-    room.users.forEach(user => {
-      this.sendToUser(user, prefix, data);
-    });
   }
